fix: move stages handlers to helpers and remove unnecessary files
